refactor: move ProfileData to the model package
