+++ app/src/main/java/com/epfl/beatlink/model/profile/ProfileData.kt
@@ -1,4 +1,4 @@
-package com.epfl.beatlink.repository.profile
+package com.epfl.beatlink.model.profile
 
 import android.net.Uri
