+++ app/src/main/java/com/epfl/beatlink/model/profile/ProfileRepository.kt
@@ -1,7 +1,5 @@
 package com.epfl.beatlink.model.profile
 
-import com.epfl.beatlink.repository.profile.ProfileData
-
 interface ProfileRepository {
   fun getUserId(): String?
