+++ settings.gradle.kts
@@ -19,7 +19,7 @@ dependencyResolutionManagement {
     }
 }
 
-rootProject.name = "SampleApp"
+rootProject.name = "BeatLink"
 include(":app")
